requirements.txt
requests
google-generativeai
python-multipart